+++ src/test/java/org/scion/jpan/api/ScmpSenderTest.java
@@ -23,8 +23,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Consumer;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 import org.junit.jupiter.api.*;
 import org.scion.jpan.*;